// src/app/features/demande_decaissement/models/entete-demande.model.ts
export interface EnteteDemande {
  iddemande?: string;
  codedemande?: string;
  iddemandeur?: string;
  typedemande?: string;
  libelledemande?: string;
  datedemande?: string; // ISO
  decaisse?: number;
  solde?: number;
  statut?: string;
  idcircuit?: string;
  idsociete?: string;
  idsite?: string;
  iddepartement?: string;
  iddevise?: string;
  createdat?: string;
  createdby?: string;
  updatedat?: string;
  updatedby?: string;
}