// src/app/features/demande_decaissement/models/ligne-demande.model.ts
export interface LigneDemande {
  idlignedemande?: string;
  iddemande: string;
  numligne?: number;
  libellelignedemande?: string;
  montantdemande?: number;
  idnature?: string;
  idbudget?: string;
  idcentre?: string;
  idsociete?: string;
  idsite?: string;
  createdat?: string;
  createdby?: string;
  updatedat?: string;
  updatedby?: string;
}